added more filters

+++ src/idpconfgen/libs/libdownload.py
@@ -3,9 +3,11 @@ import traceback
 import urllib.request
 from contextlib import contextmanager
 
+import numpy as np
+
 from idpconfgen import Path, log
 from idpconfgen.core import exceptions as EXCPTS
-from idpconfgen.libs.libstructure import Structure, col_altLoc
+from idpconfgen.libs.libstructure import Structure, col_altLoc, write_PDB, structure_to_pdb, col_resSeq, col_resName, col_element
 from idpconfgen.logger import S
 
 
@@ -17,9 +19,9 @@ POSSIBLELINKS = [
     ]
 
 
-def download_structure(pdbid, folder='', record_name='ATOM'):
+def download_structure(pdbid, **kwargs):
     """
-    Download a PDB/CIF structure.
+    Download a PDB/CIF structure chains.
 
     Parameters
     ----------
@@ -27,30 +29,39 @@ def download_structure(pdbid, folder='', record_name='ATOM'):
         0-indexed, the structure ID at RCSB.org;
         1-indexed, a list of the chains to download.
 
-    folder : str or Path, optional
-        The destination folder. Default to current working directory.
-
-    record_name : str or tuple
-        The RECORD name lines to save.
+    **kwargs : as for :func:`save_structure_chains_and_segments`.
     """
     pdbname = pdbid[0]
     chains = pdbid[1]
 
-    possible_links = (l.format(pdbname) for l in POSSIBLELINKS)
+    downloaded_data = fetch_pdb_id_from_RCSB(pdbname)
+    save_structure_chains_and_segments(
+        downloaded_data,
+        pdbname,
+        chains=chains,
+        **kwargs,
+        )
 
-    for weblink in possible_links:
-        try:
-            response = urllib.request.urlopen(weblink)
-            downloaded_data = response.read()
-            break
-        except urllib.error.HTTPError:
-            continue
-        except (AttributeError, UnboundLocalError):  # response is None
-            log.error(S(f'Download {pdbname} failed to read data.'))
-            continue
-    else:
-        log.error(S(f'Failed to download {pdbname}'))
-        return
+
+# considers solvent and DNA/RNA
+# http://www.wwpdb.org/documentation/file-format-content/format33/sect4.html#HET
+_discarded_residues = ('HOH', 'SOL', 'I', 'C', 'G', 'A',
+    'U', 'I', 'DC', 'DG', 'DA', 'DU', 'DT', 'DI', 'N')
+
+_allowed_elements = ('C', 'O', 'N', 'H', 'S', 'Se', 'D')
+
+
+def save_structure_chains_and_segments(
+        downloaded_data,
+        pdbname,
+        chains=None,
+        record_name='ATOM',
+        altlocs=('A', '', ' '),
+        folder='',
+        ):
+
+    _DR = _discarded_residues
+    _AE = _allowed_elements
 
     pdbdata = Structure(downloaded_data)
     pdbdata.build()
@@ -58,29 +69,53 @@ def download_structure(pdbid, folder='', record_name='ATOM'):
     chains = chains or pdbdata.chain_set
 
     pdbdata.add_filter_record_name(record_name)
-    pdbdata.add_filter(
-        lambda x: x[col_altLoc] in ('A', '', ' ')
-        )
+    pdbdata.add_filter(lambda x: x[col_resName] not in _DR)
+    pdbdata.add_filter(lambda x: x[col_element] in _AE)
+    pdbdata.add_filter(lambda x: x[col_altLoc] in altlocs)
 
     for chain in chains:
         pdbdata.add_filter_chain(chain)
 
-        #
+        # passar esto a una function
         pdbsegs = pdbdata.residue_segments
-        if len(pdbsegs) > 1:
-            for i, segment in enumerate(pdbsegs):
-                fout_seg = Path(folder, f'{pdbname}_{chain}_seg{i}.pdb}')
-                with try_to_write(downloaded_data, fout_seg):
-                    write_PDB(segment)
 
+        # more than one residue
+        valid_segments = filter(
+            lambda x: set(line[col_resSeq] for line in x),
+            pdbsegs,
+            )
+
+        if len(pdbsegs) > 1:
+            for i, segment in enumerate(valid_segments):
+                fout_seg = Path(folder, f'{pdbname}_{chain}_seg{i}.pdb')
+                with try_to_write(downloaded_data, fout_seg):
+                    # because segments are pure arrays
+                    # and not Structure objects
+                    write_PDB(structure_to_pdb(segment), fout_seg)
         #
+
         else:
             fout = Path(folder, f'{pdbname}_{chain}.pdb')
             with try_to_write(downloaded_data, fout):
                 pdbdata.write_PDB(fout)
 
-
         pdbdata.pop_last_filter()
+
+
+def fetch_pdb_id_from_RCSB(pdbid):
+    possible_links = (l.format(pdbid) for l in POSSIBLELINKS)
+
+    for weblink in possible_links:
+        try:
+            response = urllib.request.urlopen(weblink)
+            return response.read()
+        except urllib.error.HTTPError:
+            continue
+        except (AttributeError, UnboundLocalError):  # response is None
+            #log.error(S(f'Download {pdbid} failed to read data.'))
+            continue
+    else:
+        raise IOError(f'Failed to download {pdbid}')
 
 
 @contextmanager
@@ -94,7 +129,10 @@ def try_to_write(data, fout):
         erp = Path(fout.myparents(), 'errors')
         erp.mkdir(parents=True, exist_ok=True)
         p = Path(erp, fout.stem).with_suffix('.structure')
-        try:
-            p.write_bytes(data)
-        except TypeError:
-            p.write_text(data)
+        if p.exists():
+            return
+        else:
+            try:
+                p.write_bytes(data)
+            except TypeError:
+                p.write_text(data)
